move mail, image_file module under utils from apis directory.
