+++ django_app/utils/image_file/resizing_image.py
